@@ -12,7 +12,9 @@
         "barryvdh/laravel-dompdf": "^0.8.3",
         "fideloper/proxy": "^4.4",
         "intervention/image": "^2.4",
-        "laravel/framework": "^6.20.26",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^6.3.1|^7.0.1",
+        "laravel/framework": "^7.29",
         "laravel/tinker": "^2.5",
         "paypal/rest-api-sdk-php": "^1.14",
         "stripe/stripe-php": "^6.19"
@@ -20,10 +22,10 @@
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.1",
         "beyondcode/laravel-dump-server": "^1.0",
-        "facade/ignition": "^1.16.15",
+        "facade/ignition": "^2.0",
         "fakerphp/faker": "^1.9.1",
-        "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^3.0",
+        "mockery/mockery": "^1.3.1",
+        "nunomaduro/collision": "^4.3",
         "phpunit/phpunit": "^8.5.8|^9.3.3"
     },
     "config": {
