@@ -3,12 +3,16 @@
 namespace App\Http\Controllers\Auth;
 
 use App\User;
+use App\Models\Club;
+use App\Models\Permisologia\Role;
+use Illuminate\Support\Facades\Mail;
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\Foundation\Auth\RegistersUsers;
-use App\Models\Permisologia\Role;
-use App\Models\Club;
+// use Illuminate\Foundation\Auth\RegistersUsers;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Auth\Events\Registered;
 
 class RegisterController extends Controller
 {
@@ -23,7 +27,7 @@ class RegisterController extends Controller
     |
     */
 
-    use RegistersUsers;
+    // use RegistersUsers;
 
     /**
      * Where to redirect users after registration.
@@ -50,7 +54,7 @@ class RegisterController extends Controller
     public function showRegistrationForm()
     {
         $clubs = Club::get(['id', 'name', 'population', 'province']);
-        $roles = Role::whereIn('id', [2,3])->get(['id', 'name']);
+        $roles = Role::whereIn('id', [2, 3])->get(['id', 'name']);
         return view('auth.register', compact('clubs', 'roles'));
     }
 
@@ -64,9 +68,9 @@ class RegisterController extends Controller
     {
         return Validator::make($data, [
             'email'     => 'required|email|min:8|max:35|unique:users',/*DomainValid*/
-            'last_name' => 'required|alfa_space|min:3|max:50',
-            'user'      => 'required|alfa_space|min:3|max:20|unique:users',
-            'name'      => 'required|alfa_space|min:3|max:50',
+            'last_name' => 'required|string|min:3|max:50',
+            'user'      => 'required|string|min:3|max:20|unique:users',
+            'name'      => 'required|string|min:3|max:50',
             'num_id'    => 'required|string|min:6|max:12|exr_ced|unique:users',/*|numeric|digits_between:6,8*/
             'club_id'   => 'nullable|numeric',
             'phone'     => 'nullable|numeric',
@@ -74,7 +78,7 @@ class RegisterController extends Controller
             'sex'     => 'required|numeric',
             'web'     => 'nullable|string',
             'password'  => 'required|string|min:6|max:20|confirmed',
-        ],[],[
+        ], [], [
             'email'     => 'correo',
             'last_name' => 'apellido',
             'user'      => 'nombre de usuario',
@@ -96,12 +100,67 @@ class RegisterController extends Controller
      */
     protected function create(array $data)
     {
+        $data['user'] = $data['name'];
         $data['password'] = Hash::make($data['password']);
         $data['confirm'] = str_replace('/', '', Hash::make(Hash::make(env('APP_KEY')) . $data['password'] . '-rs'));
         $user = User::create($data);
-        \Mail::to($user->email)->send(new \App\Mail\Welcome($user));
+        Mail::to($user->email)->send(new \App\Mail\Welcome($user));
         $user->roles()->attach($data['rol']);
         $user->assignPermissionsOneUser($data['rol']);
         return $user;
+    }
+
+    /**
+     * Handle a registration request for the application.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function register(Request $request)
+    {
+        $this->validator($request->all())->validate();
+
+        event(new Registered($user = $this->create($request->all())));
+
+        $this->guard()->login($user);
+
+        return $this->registered($request, $user)
+            ?: redirect($this->redirectPath());
+    }
+
+    /**
+     * Get the guard to be used during registration.
+     *
+     * @return \Illuminate\Contracts\Auth\StatefulGuard
+     */
+    protected function guard()
+    {
+        return Auth::guard();
+    }
+
+    /**
+     * The user has been registered.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  mixed  $user
+     * @return mixed
+     */
+    protected function registered(Request $request, $user)
+    {
+        //
+    }
+
+    /**
+     * Get the post register / login redirect path.
+     *
+     * @return string
+     */
+    public function redirectPath()
+    {
+        if (method_exists($this, 'redirectTo')) {
+            return $this->redirectTo();
+        }
+
+        return property_exists($this, 'redirectTo') ? $this->redirectTo : '/home';
     }
 }
