@@ -4,7 +4,7 @@ namespace App\Providers;
 
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Schema;
-use Carbon;
+use Carbon\Carbon;
 
 class AppServiceProvider extends ServiceProvider
 {
